Updated SDL to SDL3 now that it is in the Vulkan SDK

+++ ChangeMe/src/camera/arcball_camera.cpp
@@ -1,7 +1,7 @@
 #include "arcball_camera.hpp"
 
 #include <algorithm>
-#include <SDL2/SDL_mouse.h>
+#include <SDL3/SDL_mouse.h>
 
 ArcballCamera::ArcballCamera(const glm::vec3 p_Target, const float p_DistanceToTarget, const float p_Fov, const float p_Near, const float p_Far)
     : PerspectiveCamera(glm::vec3{}, glm::vec3{0.0f, 0.0f, 1.0f}, glm::vec3{0.0f, 1.0f, 0.0f}, p_Fov, p_Near, p_Far)
@@ -16,12 +16,12 @@ void ArcballCamera::setTarget(const glm::vec3 p_Target)
     setPosition(m_Target - m_Front * m_DistanceToTarget);
 }
 
-void ArcballCamera::mouseMoved(const int32_t p_RelX, const int32_t p_RelY)
+void ArcballCamera::mouseMoved(const float p_RelX, const float p_RelY)
 {
     if (m_LeftMousePressed)
     {
-        const float l_Dx = static_cast<float>(p_RelX) * m_RotatingSpeed;
-        const float l_Dy = static_cast<float>(p_RelY) * m_RotatingSpeed;
+        const float l_Dx = p_RelX * m_RotatingSpeed;
+        const float l_Dy = p_RelY * m_RotatingSpeed;
 
         m_Yaw += l_Dx;
         m_Pitch += l_Dy;
@@ -38,8 +38,8 @@ void ArcballCamera::mouseMoved(const int32_t p_RelX, const int32_t p_RelY)
     }
     else if (m_RightMousePressed)
     {
-        const float l_Dx = static_cast<float>(p_RelX);
-        const float l_Dy = static_cast<float>(p_RelY);
+        const float l_Dx = p_RelX;
+        const float l_Dy = p_RelY;
 
         const glm::vec3 l_Delta = (m_Right * -l_Dx + m_Up * -l_Dy) * (m_MovingSpeed * 0.01f);
 
@@ -63,9 +63,9 @@ void ArcballCamera::mouseButtonReleased(const uint32_t p_Button)
         m_RightMousePressed = false;
 }
 
-void ArcballCamera::mouseScrolled(const int32_t p_Y)
+void ArcballCamera::mouseScrolled(const float p_Y)
 {
-    m_DistanceToTarget -= static_cast<float>(p_Y);
+    m_DistanceToTarget -= p_Y;
     m_DistanceToTarget = std::max(m_DistanceToTarget, 1.0f);
     setPosition(m_Target - m_Front * m_DistanceToTarget);
 }
