+++ ChangeMe/src/camera/flight_camera.hpp
@@ -9,11 +9,11 @@ class FlightCamera final : public PerspectiveCamera
 public:
 	FlightCamera(glm::vec3 p_Pos, glm::vec3 p_Dir, glm::vec3 p_Up, float p_Fov = 70.0f, float p_Near = 0.1f, float p_Far = 1000.0f);
 
-	void mouseMoved(int32_t p_RelX, int32_t p_RelY) override;
+	void mouseMoved(float p_RelX, float p_RelY) override;
 	void keyPressed(uint32_t p_Key) override;
 	void keyReleased(uint32_t p_Key) override;
 	void updateEvents(float p_Delta) override;
-    void mouseScrolled(int32_t p_Y) override;
+    void mouseScrolled(float p_Y) override;
 
 private:
     float m_MovingSpeed = 10.f;
