+++ ChangeMe/src/camera/ortho_controller_camera.hpp
@@ -7,10 +7,10 @@ public:
     OrthoControllerCamera(glm::vec3 p_Pos, glm::vec3 p_Dir, glm::vec3 p_Up, glm::vec2 p_XBounds, glm::vec2 p_YBounds, float p_Near = 0.1f, float p_Far = 1000.0f);
 
 private:
-    void mouseMoved(int32_t p_RelX, int32_t p_RelY) override;
+    void mouseMoved(float p_RelX, float p_RelY) override;
     void mouseButtonPressed(uint32_t p_Button) override;
     void mouseButtonReleased(uint32_t p_Button) override;
-    void mouseScrolled(int32_t p_Y) override;
+    void mouseScrolled(float p_Y) override;
 
 private:
     void recalculateUnitsPerPixel();
